move common items to common_ast
